Relocate folders and files
