+++ src/App/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 
-import Header from "../Header/Header";
+import Header from "../post/components/Header/Header";
 import "./App.css";
 
 const App = (): React.ReactElement => {
