+++ src/post/components/Card/Card.css
@@ -3,7 +3,6 @@
   border: 2px solid black;
   border-radius: 25px;
   max-width: 100%;
-  margin-top: 40px;
 
   @media (width >= 600px) {
     max-width: 70%;
