+++ src/post/components/Header/Header.css
@@ -1,4 +1,5 @@
 .main-header {
+  margin-bottom: 40px;
   display: flex;
   flex-direction: column;
   align-items: center;
