+++ src/router/pages/NotFoundPage/NotFoundPage.css
@@ -1,5 +1,5 @@
 .error {
   color: #ff0000;
-  font-size: 5rem;
+  font-size: 64px;
   margin: 100px auto;
 }
